+++ target-gen/src/generate.rs
@@ -1,17 +1,21 @@
-use std::fs::{self};
-use std::io::Read;
-use std::path::Path;
-
 use anyhow::{anyhow, bail, Context, Error, Result};
 use cmsis_pack::pdsc::{Core, Device, Package, Processor};
 use cmsis_pack::{pack_index::PdscRef, utils::FromElem};
 use futures::StreamExt;
-use probe_rs::config::{
-    Chip, ChipFamily, Core as ProbeCore, GenericRegion, MemoryRegion, NvmRegion, RamRegion,
-    RawFlashAlgorithm,
+use probe_rs::{
+    config::{
+        Chip, ChipFamily, Core as ProbeCore, GenericRegion, MemoryRegion, NvmRegion, RamRegion,
+        RawFlashAlgorithm,
+    },
+    flashing::FlashAlgorithm,
+    Architecture, CoreType,
 };
-use probe_rs::{Architecture, CoreType};
 use probe_rs_target::{ArmCoreAccessOptions, CoreAccessOptions, RiscvCoreAccessOptions};
+use std::{
+    fs::{self},
+    io::Read,
+    path::Path,
+};
 use tokio::runtime::Builder;
 
 pub(crate) enum Kind<'a, T>
@@ -81,7 +85,7 @@ where
             .algorithms
             .iter()
             .map(|flash_algorithm| {
-                let algo = match &mut kind {
+                let mut algo = match &mut kind {
                     Kind::Archive(archive) => crate::parser::extract_flash_algo(
                         archive.by_name(&flash_algorithm.file_name.as_path().to_string_lossy())?,
                         &flash_algorithm.file_name,
@@ -95,6 +99,15 @@ where
                         false, // Algorithms from CMSIS-Pack files are position independent
                     ),
                 }?;
+
+                // If the algo specifies `RAMstart` and/or `RAMsize` fields, then use them.
+                // - See https://open-cmsis-pack.github.io/Open-CMSIS-Pack-Spec/main/html/pdsc_family_pg.html#element_algorithm for more information.
+                algo.load_address = flash_algorithm.ram_start.map(|ram_start| ram_start + FlashAlgorithm::get_max_algorithm_header_size());
+                if let Some(stack_size) = flash_algorithm.ram_size {
+                     algo.stack_size = Some(stack_size.try_into().map_err(|data_conversion_error|
+                        anyhow!("Algorithm requires a stack size of  '{:?}' : {data_conversion_error:?}", flash_algorithm.ram_size)
+                    )?);
+                }
 
                 // We add this algo directly to the algos of the family if it's not already added.
                 // Make sure we never add an algo twice to save file size.
@@ -447,13 +460,9 @@ struct DeviceMemory {
     name: String,
 }
 
-// From PR's https://github.com/probe-rs/target-gen/pull/20 and https://github.com/probe-rs/target-gen/pull/25:
-// TODO: What is the logic that justifies PR#20 selecting the largest memory? Shouldn't we match flash algo's with RAM based on load_address?
-// Flash and RAM regions are not guaranteed to be sorted in the PDSC file, so we:
-// - Sort them here.
-// - Merge contiguous regions.
-// Update: For multiple cores, we have to take processor access into account during this merge.
-/// Sorts the memory regions in the package and merges contiguous regions with the same attributes.
+/// Extracts the memory regions in the package.
+/// The new memory regions are sorted by memory type, then by boot memory, then by start address,
+/// with correctly assigned cores/processor names.
 pub(crate) fn get_mem_map(device: &Device) -> Vec<MemoryRegion> {
     let mut device_memories: Vec<DeviceMemory> = device
         .memories
@@ -479,36 +488,8 @@ pub(crate) fn get_mem_map(device: &Device) -> Vec<MemoryRegion> {
         })
         .collect();
 
-    // Merge memory regions with the same attributes.
-    if device_memories.len() > 1 {
-        // Sort by memory type, then by processor name, then by boot memory, then by start address.
-        device_memories.sort();
-
-        let mut merged: Vec<DeviceMemory> = Vec::new();
-        let mut cur = device_memories.first().cloned().unwrap();
-        for region in device_memories.iter().skip(1) {
-            if region.is_boot_memory == cur.is_boot_memory && region.memory_start == cur.memory_end
-            {
-                // Merge with previous region.
-                cur.memory_end = region.memory_end;
-                cur.name = format!("{} + {}", cur.name, region.name);
-            } else {
-                merged.push(cur);
-                cur = region.clone();
-            }
-        }
-        merged.push(cur);
-        device_memories = merged;
-    }
-
-    // Finally, sort so that the LARGEST contiguous region is first for each core.
-    device_memories.sort_by_cached_key(|region| {
-        (
-            region.memory_type.clone(),
-            region.p_name.clone(),
-            (region.memory_start as i128 - region.memory_end as i128),
-        )
-    });
+    // Sort by memory type, then by processor name, then by boot memory, then by start address.
+    device_memories.sort();
 
     // Convert DeviceMemory's to MemoryRegion's, and assign cores to shared reqions.
     let mut mem_map = vec![];